Add Icon Rotation Animation and Icon Script
This commit adds the icon rotation animation and icon script. This
animation is added to the subject icons that will be placed in the
scene. The animation will be triggered using the Icon script that is
called when the player places the cardboard reticle over the icon. The
associated script will also stop the animation when the reticle leaves
the icon.