Add Teleport Script
This commit adds the teleport script. This script is called when the
player places the cardboard reticle over a subject icon and presses the
cardboard trigger. When the TeleportToMarker function is called the
player’s position is changed to the position of the selected subject
icon with a slight offset so that the player is standing next to the
desk beneath the icon.