Add InfoDisplay Prefab and Class
The InfoDisplay is the information window that is displayed when a
subject is selected. The InfoDisplay contains four elements: a photo, a
photo caption, a title, and a description. This information will be
passed in from the subject icon, which will be updated to include
public variables for these elements.