Update Icon Class
This commit updates the icon class to include four public variables
that will contain the information to be passed into the InfoDisplay.
This update also adds three new methods: ShowInfo, which displays the
InfoDisplay; EnablePaper, which turns on the collider for the
associated subject paper; and DisablePaper, which disables the collider
for the associated subject paper.