Update Teleporter Class
This commit adds references to the GameManager which is used to store a
reference to the selected merker/icon. The camera variable has also
been changed to type GvrViewer so that when the position of the camera
changes, the parent object (GvrViewerMain) is moved instead of the
camera itself. This was done to address some minor motion issues when
moving the camera directly.