Update Classroom
This commit adds the start screen and wires up all of the subject icons
and papers to the info display. The info display is instantiated when
as subject paper is clicked on. Clicking on the info display window
will close the window so that another subject icon can be selected.