Update GameManager Class
This commit adds the references to the ToolTip class and prefab. Two
additional functions were added to handle the displaying of the tooltip
as well as closing.

When the player hovers the reticle over the subject paper the tooltip
is displayed (instantiated). When the player moves the reticle away
from the subject paper the tooltip is closed (destroyed).